requirements.txt
numpy
anndata
packaging
wheel
numba
scanpy
tables
paste-bio
Cython
POT
